SSPCPP-691 replace ignoreCase with caseSensitive
https://issues.shibboleth.net/jira/browse/SSPCPP-691

This commit creates a new helper function to do the right thing
with respect to the (soon to be deprecated) "ignoreCase" Attribute
and the (soon to be widespread) "caseSensitive" Attribute.

In the rare cases of deprecated or broken function being encountered
we grow a logging::Category and use it to log.