CPPXT-110 OpenSSL 1.1 makes X509_STORE_CTX and X509_CRL opaque
https://issues.shibboleth.net/jira/browse/CPPXT-110

Where possible we use the setters and getters.

In some cases the setters and getters had their names changed and so
hide this inside the new header file security/impl/OpenSSLSupport.h
We are keeping the header file inside impl for now.

In PKIXPathValidator we used to have one on the stack. Instead
security/impl/OpenSSLSupport (.h and .cpp) invent a new RAII class to
abstract the allocation away.  The class also has a couple of methods
which hide the changes to setter/getter names
(X509_STORE_CTX_set0_trusted_stack and X509_STORE_CTX_get0_chain)

Also in PKIXPathValidator a series of macros around X509_CRL have
become a function without a const parameter so we conditionalize
a local variables 'const' nesss.