Unwind previous.
Managed to get myself into a fankle and whilst I create
a branch and then make changes I pushed the changes and
then made the remote branch.  Sigh.

This leaves mainline in a precarious position, so this backs
it all out.

I'll commit the inverse change to the new branch and from then on
changes should be as I wanted..

Maybe