CPPXT-110 Add test path for RSA loadXXXBigNums
https://issues.shibboleth.net/jira/browse/CPPXT-110

This completes the code coervage tests for the OpenSSLCryptoKeyRSA
changes for OpenSSL 1.1

Add a test function to force the load of the RSA key from the
<ds:RSAKeyValue>, then cast it and the credential into the
base OpenSSL type.

Throw various OpenSSL incantations at these to extract the OpenSSL
struct and compare the two.  This will need work when we move to
OpenSSL 1.1 (since the structures become opaque).