CPPXT-106 Apache Deadlock in load thread/lock thread
https://issues.shibboleth.net/jira/browse/CPPXT-106

Under certain cirucmstances an Apache thread can
cause us to do a Read then Write lock.  Diagnosed as being because
of the lock dance in the ReloadableXMLFile::lock() call.

Move the dance and updatingof m_filestamp to the reload thread,
all the lock api needs to do is spot that things are out of date
and kick the worker thread.  This might cause the worker to wake
up too often but this is unlikely.